Fix issue with sending BluetoothCallQualityReport to Telecom.
The negative ack count would always be zero due to the fact that both
packetsNotReceivedCount and negativeAcknowledgementCount were assigned to
the packetsNotReceivedCount property of the builder.

Modified isNullCall to call BluetoothCall#isCallNull so that the result
can be mocked out to facilitate the unit test (otherwise we'd need to be
able to return a mock android.telecom.Call which is not possible due to
the fact it is a final class).

Tag: #stability
Test: atest BluetoothInCallServiceTest
Bug: 200389658
Change-Id: I5af150fda830b036f278552782f394989586acb9