Disable the SignalStrengthUpdateRequest in Bluetooth pkg
As part of the work to replace feature "always report signal strength"
with SignalStrengthUpdateRequest, Bluetooth pkg introduced
SignalStrengthUpdateRequest in aosp/1569860. However, a crash
is observed.

To resolve the crash while keep the feature continuing, the logic
involves SignalStrengthUpdateRequest has been removed temporarily.
A follow-up CL will bring it back when the crash is fixed.

Bug: 205513835
Test: Manually test to confirm no crash
Change-Id: Ic9c4d5ec118b5d026341a4c024236a46d52b5605